+++ lib/goo/base/where.rb
@@ -169,11 +169,6 @@ module Goo
       end
 
       def all
-        if @result.nil? && @klass.inmutable? && @klass.inm_instances
-          if @pattern.nil? && @filters.nil?
-            @result = @klass.inm_instances.values
-          end
-        end
         process_query unless @result
         @result
       end
